+++ src/constellation/api_devices_list.go
@@ -44,13 +44,13 @@ func DeviceList(w http.ResponseWriter, req *http.Request) {
 	if isAdmin {
 		// If admin, get all devices
 		cursor, err := c.Find(nil, map[string]interface{}{})
-		defer cursor.Close(nil)
 		if err != nil {
 			utils.Error("DeviceList: Error fetching devices", err)
 			utils.HTTPError(w, "Error fetching devices", http.StatusInternalServerError, "DL001")
 			return
 		}
-		
+		defer cursor.Close(nil)
+
 		if err = cursor.All(nil, &devices); err != nil {
 			utils.Error("DeviceList: Error decoding devices", err)
 			utils.HTTPError(w, "Error decoding devices", http.StatusInternalServerError, "DL002")
@@ -60,13 +60,13 @@ func DeviceList(w http.ResponseWriter, req *http.Request) {
 		// If not admin, get user's devices based on their nickname
 		nickname := utils.GetAuthContext(req).Nickname
 		cursor, err := c.Find(nil, map[string]interface{}{"Nickname": nickname})
-		defer cursor.Close(nil)
 		if err != nil {
 			utils.Error("DeviceList: Error fetching devices", err)
 			utils.HTTPError(w, "Error fetching devices", http.StatusInternalServerError, "DL003")
 			return
 		}
-		
+		defer cursor.Close(nil)
+
 		if err = cursor.All(nil, &devices); err != nil {
 			utils.Error("DeviceList: Error decoding devices", err)
 			utils.HTTPError(w, "Error decoding devices", http.StatusInternalServerError, "DL004")